+++ src/main/java/com/prowidesoftware/swift/model/field/Field37U.java
@@ -27,7 +27,7 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.*;
-import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.Strings;
 
 /**
  * @deprecated This field has been removed since SRU2025
@@ -81,7 +81,7 @@ public class Field37U extends Field implements Serializable, AmountContainer {
         if (tag == null) {
             throw new IllegalArgumentException("tag cannot be null.");
         }
-        if (!StringUtils.equals(tag.getName(), "37U")) {
+        if (!Strings.CS.equals(tag.getName(), "37U")) {
             throw new IllegalArgumentException("cannot create field 37U from tag " + tag.getName()
                     + ", tagname must match the name of the field.");
         }