+++ src/main/java/com/prowidesoftware/swift/utils/MurMessageComparator.java
@@ -18,7 +18,7 @@ package com.prowidesoftware.swift.utils;
 import com.prowidesoftware.swift.model.SwiftMessage;
 import java.util.Comparator;
 import java.util.Objects;
-import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.Strings;
 
 /**
  * Compares the messages based on the MUR.
@@ -46,6 +46,6 @@ public class MurMessageComparator implements Comparator<SwiftMessage> {
     public int compare(final SwiftMessage left, final SwiftMessage right) {
         Objects.requireNonNull(left);
         Objects.requireNonNull(right);
-        return StringUtils.equals(left.getMUR(), right.getMUR()) ? 0 : 1;
+        return Strings.CS.equals(left.getMUR(), right.getMUR()) ? 0 : 1;
     }
 }